refactor: 인터페이스 및 레포지토리 위치 변경
