+++ src/v1/cache/interfaces/base.cache.interface.ts
