+++ src/v1/repositories/interfaces/base.repository.interface.ts
