+++ src/v1/repositories/interfaces/user.repository.interface.ts
